// src/api/generated/schemas/getBillingRequests200DataItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Elysia Documentation
 * Development documentation
 * OpenAPI spec version: 0.0.0
 */
import type { GetBillingRequests200DataItemStatus } from './getBillingRequests200DataItemStatus';
import type { GetBillingRequests200DataItemContactChannel } from './getBillingRequests200DataItemContactChannel';
import type { GetBillingRequests200DataItemNote } from './getBillingRequests200DataItemNote';
import type { GetBillingRequests200DataItemRejectionReason } from './getBillingRequests200DataItemRejectionReason';
import type { GetBillingRequests200DataItemCreatedAt } from './getBillingRequests200DataItemCreatedAt';
import type { GetBillingRequests200DataItemReviewedAt } from './getBillingRequests200DataItemReviewedAt';
import type { GetBillingRequests200DataItemCancelledAt } from './getBillingRequests200DataItemCancelledAt';
import type { GetBillingRequests200DataItemUser } from './getBillingRequests200DataItemUser';
import type { GetBillingRequests200DataItemReviewedBy } from './getBillingRequests200DataItemReviewedBy';

export type GetBillingRequests200DataItem = {
  id: string;
  lenses: number;
  unitPriceUsd: string;
  totalUsd: string;
  status: GetBillingRequests200DataItemStatus;
  contactChannel: GetBillingRequests200DataItemContactChannel;
  contactHandle: string;
  /** @nullable */
  note: GetBillingRequests200DataItemNote;
  /** @nullable */
  rejectionReason: GetBillingRequests200DataItemRejectionReason;
  createdAt: GetBillingRequests200DataItemCreatedAt;
  /** @nullable */
  reviewedAt: GetBillingRequests200DataItemReviewedAt;
  /** @nullable */
  cancelledAt: GetBillingRequests200DataItemCancelledAt;
  userEmail: string;
  locale: string;
  /** @nullable */
  user: GetBillingRequests200DataItemUser;
  /** @nullable */
  reviewedBy: GetBillingRequests200DataItemReviewedBy;
};