// src/api/generated/schemas/getBillingRequestsStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Elysia Documentation
 * Development documentation
 * OpenAPI spec version: 0.0.0
 */

export type GetBillingRequestsStatus = typeof GetBillingRequestsStatus[keyof typeof GetBillingRequestsStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetBillingRequestsStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  CANCELLED: 'CANCELLED',
} as const;